requirements.txt
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.4
python-dotenv==1.0.1
django-cors-headers==4.3.1
numpy==1.26.4
pandas
scikit-learn
requests==2.31.0
python-jose==3.3.0
bcrypt==4.1.2
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1 
